Add integrity hashes, note CSP rules
Even though we're serving up scripts from our own domain and *not* a
CDN, it's very little effort to just add a hash as part of the
`<script>` tag for a bit of free security.

Additionally, this project is deployed with a CSP. Because this
repository only has JavaScript source (and doesn't deal with giving
headers), it's worth noting our expectation of compatibility.

For reference, here's the full CSP I serve with this project
(newlines added for readability):

    default-src 'none';
    base-uri 'self';
    connect-src 'self' https://api.dcain.me;
    form-action 'self';
    frame-ancestors 'none';
    img-src 'self' data: https://secure.syndetics.com;
    script-src 'self';
    style-src 'self'

This can be easily tested locally with `serve`:

1. First, configure serve to give the headers:

    ```json
    # public/serve.json
    {
      "headers": [
        {
          "source": "**/*",
          "headers": [
            {
              "key": "Content-Security-Policy",
              "value": "base-uri 'self'; ... style-src 'self'"
            }
          ]
        }
      ]
    }
    ```

2. Build bundle:

    ```bash
     npm run build
    ```

3. Serve, check console for errors:

    ```bash
    npx serve dist
    ```