Show timestamp when data last cached
This surfaces the *information* to end users, but doesn't yet provide
any tools to do anything with it.

The next step will be to give a refresh button that allows users to skip
the cache entirely.

I might play with the one-minute cutoff slightly, though it's a sensible
start - calling the endpoint any more than once per minute is likely not
needed in any but the rarest cases. Additionally, it's not really super
helpful to show to users that data was cached "seconds ago."