Update SFPL branches to reflect COVID changes
While the chief motivating reason for this repository is a thing of the
past (I no longer can just walk into the main branch of the SFPL to pick
out books), I'm still interested in using this tool to find books in the
SFPL system.

From SPFL's COVID page (https://sfpl.org/about/coronavirus):

> Branch Alert
> ============
> **Available**:
>
> These locations [Main, Anza, Bayview Bookmobile, Chinatown, Eureka
> Valley, Excelsior, Marina, Merced, Mission Bay, Mission Bookmobile, OMI
> Bookmobile, Ortega, Treasure Island Bookmobile, Visitacion Valley] are
> open for SFPL-to-Go front door holds pickup service
>
> **Closed Now**:
>
> Most SFPL locations remain closed. Please check back for updated open
> information.

The next iteration will probably be to start showing which books can be
put on hold using SFPL-to-Go.