Save the last-used form contents
When biblio.dcain.me is loaded for the first time, it defaults for my
user ID and my local branch. This is helpful for me (since I just click
the button when I first load the page).

So that others may enjoy a similarly convenient workflow, save their
form chaices (with some resilience against future changes).